KNN/requirements.txt
scikit-learn
numpy
pandas
matplotlib
opencv-python
pillow